% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine_similarity.R
\name{cosine_similarity}
\alias{cosine_similarity}
\title{Cosine Similarity}
\usage{
cosine_similarity(x, y, verbose = FALSE, ...)
}
\arguments{
\item{x}{A numeric vector containing no NA}

\item{y}{A numeric vector (containing no NA) of the same size as \code{x}}

\item{verbose}{A logic value (TRUE or FALSE, default is FALSE). When it is TRUE, the function will show calculation steps.}

\item{...}{Other arguments to be passed to methods, that are not currently used.}
}
\value{
A number that is the cosine value of the angle between 2 same-size vectors \verb{x' and }y`. Output range is between -1 and 1.
}
\description{
Cosine Similarity
}
\examples{
cosine_similarity(c(1,0), c(1,1))
}
